логика рендера события и компонент форма события
