+++ src/components/menu.js
@@ -23,6 +23,10 @@ class Menu {
     }
     return this._element;
   }
+
+  removeElement() {
+    this._element = null;
+  }
 }
 
 export default Menu;
