+++ src/components/sort.js
@@ -62,6 +62,10 @@ class Sort {
     }
     return this._element;
   }
+
+  removeElement() {
+    this._element = null;
+  }
 }
 
 export default Sort;
