+++ src/main.js
@@ -1,12 +1,12 @@
-import TripEvent from './components/trip-event.js';
 
+import getTripEventData from './components/event-data';
 import Menu from './components/menu.js';
 import TripInfoData from './components/trip-info.js';
 import Filter from './components/filter.js';
 import Sort from './components/sort.js';
 import TripDay from './components/trip-day.js';
 
-import {render, unrender, createElement, Position} from './components/utils.js';
+import {render, createElement, Position} from './components/utils.js';
 
 
 const DAYS_COUNT = 3;
@@ -20,8 +20,7 @@ const EVENTS_IN_DAY_COUNT = 4;
 
 const init = () => {
   // готовим исходные данные, массив дней, в каждом дне массив событий
-  const tripDays = new Array(DAYS_COUNT).fill(``).map(() => new Array(EVENTS_IN_DAY_COUNT).fill(``).map(() => new TripEvent()));
-
+  const tripDays = new Array(DAYS_COUNT).fill(``).map(() => new Array(EVENTS_IN_DAY_COUNT).fill(``).map(() => getTripEventData()));
   // инфа о поездке
   const tripInfoData = new TripInfoData(tripDays);
   const tripInfoContainer = document.querySelector(`.trip-main__trip-info`);
